OXDEV-3103 deprecate BeforeAuthorizationEvent
